jsHistory, jsReplace

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -2,8 +2,11 @@ package com.example.demo.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
@@ -14,87 +17,113 @@ import com.example.demo.vo.ResultData;
 
 @Controller
 public class UsrArticleController {
-	
+
 	private ArticleService articleService;
-	
+
 	@Autowired
 	public UsrArticleController(ArticleService articleService) {
 		this.articleService = articleService;
 	}
-	
+
 	// 액션 메서드
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
-	public ResultData<Article> doAdd(String title, String body) {
-		
+	public ResultData<Article> doAdd(HttpSession session, String title, String body) {
+
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
+
 		if (Util.empty(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요");
 		}
-		
+
 		if (Util.empty(body)) {
 			return ResultData.from("F-2", "내용을 입력해주세요");
 		}
-		
-		articleService.writeArticle(title, body);
-		
+
+		articleService.writeArticle((int) session.getAttribute("loginedMemberId"), title, body);
+
 		int id = articleService.getLastInsertId();
-		
-		return ResultData.from("S-1", Util.f("%d번 게시글이 생성되었습니다", id), articleService.getArticleById(id));
+
+		return ResultData.from("S-1", Util.f("%d번 게시글이 생성되었습니다", id), "article", articleService.getArticleById(id));
 	}
 
-	@RequestMapping("/usr/article/getArticles")
-	@ResponseBody
-	public ResultData<List<Article>> getArticles() {
-		
+	@RequestMapping("/usr/article/list")
+	public String showList(Model model) {
+
 		List<Article> articles = articleService.getArticles();
+
+		model.addAttribute("articles", articles);
 		
-		if (articles.size() == 0) {
-			return ResultData.from("F-1", "게시물이 존재하지 않습니다");
+		return "usr/article/list";
+	}
+
+	@RequestMapping("/usr/article/detail")
+	public String showDetail(Model model, HttpSession session, int id) {
+
+		int loginedMemberId = 0;
+		
+		if(session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int) session.getAttribute("loginedMemberId");
 		}
 		
-		return ResultData.from("S-1", "게시물 리스트", articles);
+		Article article = articleService.getForPrintArticle(id);
+
+		model.addAttribute("article", article);
+		model.addAttribute("loginedMemberId", loginedMemberId);
+		
+		return "usr/article/detail";
 	}
-	
-	@RequestMapping("/usr/article/getArticle")
-	@ResponseBody
-	public ResultData<Article> getArticle(int id) {
-		
-		Article foundArticle = articleService.getArticleById(id);
-		
-		if (foundArticle == null) {
-			return ResultData.from("F-1", Util.f("%d번 게시글은 존재하지 않습니다", id));
-		}
-		
-		return ResultData.from("S-1", Util.f("%d번 게시글 입니다", id), foundArticle);
-	}
-	
+
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
-		
-		Article foundArticle = articleService.getArticleById(id);
-		
-		if (foundArticle == null) {
-			return ResultData.from("F-1", Util.f("%d번 게시글은 존재하지 않습니다", id));
+	public ResultData<Article> doModify(HttpSession session, int id, String title, String body) {
+
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
+
+		if(Util.empty(id)) {
+			return ResultData.from("F-1", "수정할 글 번호를 입력해주세요");
 		}
 		
-		articleService.modifyArticle(id, title, body);
-		
-		return ResultData.from("S-1", Util.f("%d번 게시글을 수정했습니다", id));
+		Article foundArticle = articleService.getArticleById(id);
+
+		if (foundArticle == null) {
+			return ResultData.from("F-2", Util.f("%d번 게시글은 존재하지 않습니다", id));
+		}
+
+		ResultData actorCanModifyRd = articleService.actorCanModify((int) session.getAttribute("loginedMemberId"),
+				foundArticle.getMemberId());
+
+		if (actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
+		}
+
+		return articleService.modifyArticle(id, title, body);
 	}
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
-		
-		Article foundArticle = articleService.getArticleById(id);
-		
-		if (foundArticle == null) {
-			return ResultData.from("F-1", Util.f("%d번 게시글은 존재하지 않습니다", id));
+	public String doDelete(HttpSession session, int id) {
+
+		if (session.getAttribute("loginedMemberId") == null) {
+			return Util.jsHistoryBack("로그인 후 이용해주세요");
 		}
 		
+		Article foundArticle = articleService.getArticleById(id);
+
+		if (foundArticle == null) {
+			return Util.jsHistoryBack(Util.f("%d번 게시글은 존재하지 않습니다", id));
+		}
+
+		if ((int) session.getAttribute("loginedMemberId") != foundArticle.getMemberId()) {
+			return Util.jsHistoryBack("해당 게시글에 대한 권한이 없습니다");
+		}
+
 		articleService.deleteArticle(id);
-		
-		return ResultData.from("S-1", Util.f("%d번 게시글을 삭제했습니다", id));
+
+		return Util.jsReplace(Util.f("%d번 게시글을 삭제했습니다", id), "list");
 	}
 }
